Modifications to the way toasts are displayed.
Previously the toast view was added to the application key window which made
the orientation change support difficult.

With the new implementation the toast view is added, by default, to the view of
the root view controller. Or, the root view controller can supply another view
controller to act toast delegate via the `getToastController`-method from the
`RAToastControllerDelegate`-protocol.