Fixed issue with log level, and added base for `setDelegate:`.
The current implementation for getting the toast delegate is somewhat
cumbersome, it would be better to just set the delegate you'd want.

And, not like now, have to declare the delegate protocol on the root
view controller and from there retrieve the controller delegate.

The implementation for `setDelegate:` is not yet complete.